refactor system: system.go
